requirements.txt
numpy
pygame-ce
librosa
audio_metadata